Use integers for phases

+++ pttools/bubble/const.py
@@ -19,5 +19,6 @@ dxi_small: float = 1./N_XI_DEFAULT
 
 cs0: np.float_ = 1 / np.sqrt(3)  # ideal speed of sound
 cs0_2: float = 1./3  # ideal speed of sound squared
-symm_phase: float = 0.0
-brok_phase: float = 1.0
+
+symm_phase: int = 0
+brok_phase: int = 1
